refactor: update example and move it to tests source folder
